Add sortCardsByCardholderName method. Turn currentBalance, remainingLimit and earnedCashback fields into optional.
