+++ src/pages/workspace/expensifyCard/WorkspaceExpensifyCardListPage.tsx
@@ -15,8 +15,7 @@ import useLocalize from '@hooks/useLocalize';
 import usePolicy from '@hooks/usePolicy';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-import localeCompare from '@libs/LocaleCompare';
-import * as PersonalDetailsUtils from '@libs/PersonalDetailsUtils';
+import * as CardUtils from '@libs/CardUtils';
 import Navigation from '@navigation/Navigation';
 import type {FullScreenNavigatorParamList} from '@navigation/types';
 import CONST from '@src/CONST';
@@ -48,17 +47,7 @@ function WorkspaceExpensifyCardListPage({route, cardsList}: WorkspaceExpensifyCa
 
     const policyCurrency = useMemo(() => policy?.outputCurrency ?? CONST.CURRENCY.USD, [policy]);
 
-    const sortedCards = useMemo(
-        () =>
-            Object.values(cardsList ?? {}).sort((cardA: Card, cardB: Card) => {
-                const userA = personalDetails?.[cardA.accountID ?? '-1'] ?? {};
-                const userB = personalDetails?.[cardB.accountID ?? '-1'] ?? {};
-                const aName = PersonalDetailsUtils.getDisplayNameOrDefault(userA);
-                const bName = PersonalDetailsUtils.getDisplayNameOrDefault(userB);
-                return localeCompare(aName, bName);
-            }),
-        [cardsList, personalDetails],
-    );
+    const sortedCards = useMemo(() => CardUtils.sortCardsByCardholderName(cardsList, personalDetails), [cardsList, personalDetails]);
 
     const getHeaderButtons = () => (
         <View style={[styles.w100, styles.flexRow, styles.gap2, shouldUseNarrowLayout && styles.mb3]}>
