+++ src/types/onyx/ExpensifyCardSettings.ts
@@ -3,13 +3,13 @@ import type * as OnyxCommon from './OnyxCommon';
 /** Model of Expensify card settings for a workspace */
 type ExpensifyCardSettings = OnyxCommon.OnyxValueWithOfflineFeedback<{
     /** Sum of all posted Expensify Card transactions */
-    currentBalance: number;
+    currentBalance?: number;
 
     /** Remaining limit for Expensify Cards on the workspace */
-    remainingLimit: number;
+    remainingLimit?: number;
 
     /** The total amount of cash back earned thus far */
-    earnedCashback: number;
+    earnedCashback?: number;
 
     /** The date of the last settlement */
     monthlySettlementDate: Date;
